requirements.txt
huggingface_hub
python-dotenv
rich
pydantic #helps with structured tool calling
google-generativeai
openai
